+++ packages/koot/defaults/i18n.js
@@ -1,0 +1,6 @@
+/** 默认 i18n 配置 */
+module.exports = {
+    use: 'query',
+    expr: '__',
+    cookieKey: 'spLocaleId',
+};